+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/autoinstall/AutoInstallState.kt
@@ -52,6 +52,10 @@ class AutoInstallState private constructor() : Serializable {
 
     @get:Synchronized
     @set:Synchronized
+    var installQuartz: Boolean = false
+
+    @get:Synchronized
+    @set:Synchronized
     var installKotlinExtensions: Boolean = false
 
     override fun toString(): String {
@@ -66,6 +70,7 @@ class AutoInstallState private constructor() : Serializable {
             "installLog4j2=$installLog4j2, " +
             "installJdbc=$installJdbc), " +
             "installGraphql=$installGraphql), " +
+            "installQuartz=$installQuartz), " +
             "installKotlinExtensions=$installKotlinExtensions)"
     }
 
@@ -86,6 +91,7 @@ class AutoInstallState private constructor() : Serializable {
         if (installLog4j2 != other.installLog4j2) return false
         if (installJdbc != other.installJdbc) return false
         if (installGraphql != other.installGraphql) return false
+        if (installQuartz != other.installQuartz) return false
         if (installKotlinExtensions != other.installKotlinExtensions) return false
 
         return true
@@ -103,6 +109,7 @@ class AutoInstallState private constructor() : Serializable {
         result = 31 * result + installLog4j2.hashCode()
         result = 31 * result + installJdbc.hashCode()
         result = 31 * result + installGraphql.hashCode()
+        result = 31 * result + installQuartz.hashCode()
         result = 31 * result + installKotlinExtensions.hashCode()
         return result
     }