+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- Auto install `sentry-quartz` if `quartz` is installed ([#562](https://github.com/getsentry/sentry-android-gradle-plugin/pull/562))
+
 ### Internal
 
 - Change room span op name to `db.sql.room` ([#557](https://github.com/getsentry/sentry-android-gradle-plugin/pull/557))